Define Server

Define Server

Usage

gsmApp_Server(
  dfAnalyticsInput,
  dfBounds,
  dfGroups,
  dfMetrics,
  dfResults,
  fnFetchData,
  chrDomains = c("AE", "ENROLL", "LB", "PD", "SDRGCOMP", "STUDCOMP", "SUBJ", "DATACHG",
    "DATAENT", "QUERY"),
  lPlugins = NULL,
  fnServer = NULL
)

Arguments

dfAnalyticsInput

data.frame Participant-level metric data.

dfBounds

data.frame Set of predicted percentages/rates and upper- and lower-bounds across the full range of sample sizes/total exposure values for reporting.

dfGroups

data.frame Group-level metadata dictionary.

dfMetrics

data.frame Metric-specific metadata for use in charts and reporting.

dfResults

data.frame A stacked summary of analysis pipeline output. This will be filtered to cases where GroupLevel == "Site".

fnFetchData

function A function that takes a strDomain argument and optional strSiteID and strSubjectID and returns a data.frame. See sample_fnFetchData() for an example. The returned data.frame contains information about the named domain.

chrDomains

character A vector of names of domain data.frames to show in the app. Supported domains are shown in the default value.

lPlugins

list Optional list of plugins to include in the app.

fnServer

function A Shiny server function that takes arguments input, output, and session. This function will be called at the start of the main app server function.

Value

The main server function for use in a shiny app.