Participant Details server

Update Participant Details when the selected participant changes.

Usage

mod_ParticipantDetails_Server(id, fnFetchParticipantData, rctv_strSubjectID)

Arguments

id

character The id for this module, widget, or other element.

fnFetchParticipantData

function A function that takes a single strSubjectID argument such as "0001", and returns a list with components metadata and metric_data. metadata should be a named character vector (or something that can be coerced to a character vector) of information about the participant, and metric_data should be a named list of data.frames, each of which contains information related to the named metric.

rctv_strSubjectID

reactive character A shiny::reactive() object that returns the SubjectID of the selected participant.

Value

A shiny::reactive() with the ids of rows selected in the table (currently unused but implemented as a proof-of-concept for future extensions).