Plugin Wrapper Server

Plugin Wrapper Server

Usage

mod_Plugin_Server(
  lPlugin,
  id,
  lDataModel,
  l_rctvDomains,
  l_rctvDomainHashes,
  l_rctvInputs,
  chrInputNamesPretty
)

Arguments

lPlugin

list A named list with required elements meta and shiny, and optional fields spec, packages and required_inputs. Usually generated by plugin_Read().

id

character The id for this element.

lDataModel

list Named list of the standard gsm dataframes (dfAnalyticsInput, dfBounds, dfGroups, dfMetrics, and dfResults).

l_rctvDomains

list A named list of shiny::reactive() objects, each of which returns a domain dataframe.

l_rctvDomainHashes

list A named list of shiny::reactive() objects, each of which returns the rlang::hash() of a domain dataframe.

l_rctvInputs

list A named list of shiny::reactiveVal() objects, each of which returns the current value of an input or input-like variable.

chrInputNamesPretty

character A vector of names to display for missing inputs.