Read a Plugin Definition

Plugins are defined by a named list with elements strTitle (a title to show in the plugin menu), fnUI (a function or the name of a function to use as the plugin's module UI), and fnServer (a function or the name of a function to use as the plugin's module server). Use plugin_Read() to read these element definitions and any R files in the same directory.

Usage

plugin_Read(strPath)

Arguments

strPath

The directory that contains the plugin.

Value

A list with elements strTitle, fnUi, fnServer, and (optionally) lConfig, read from the YAML file in strPath. As a side effect, any R files in strPath are also loaded using source().

Examples

aePlugin <- plugin_Read(system.file("plugins", "AE", package = "gsm.app"))
aePlugin
#> $meta
#> $meta$Type
#> [1] "Plugin"
#> 
#> $meta$ID
#> [1] "AE"
#> 
#> $meta$Name
#> [1] "Adverse Events"
#> 
#> 
#> $shiny
#> $shiny$UI
#> [1] "mod_AE_UI"
#> 
#> $shiny$Server
#> [1] "mod_AE_Server"
#> 
#> 
#> $domains
#> [1] "AE"   "SUBJ"
#> 
#> $packages
#> $packages[[1]]
#> $packages[[1]]$name
#> [1] "ggplot2"
#> 
#> 
#> $packages[[2]]
#> $packages[[2]]$name
#> [1] "safetyCharts"
#> 
#> $packages[[2]]$remote
#> [1] "url::https://safetygraphics.r-universe.dev/src/contrib/safetyCharts_0.4.0.tar.gz"
#> 
#> 
#>