Run App

Run App

Usage

run_gsm_app(
  dfResults,
  dfGroups,
  dfMetrics,
  dfBounds,
  dfAnalyticsInput,
  fnFetchParticipantData,
  strTitle = "GSM Deep Dive"
)

Arguments

dfResults

data.frame A stacked summary of analysis pipeline output. This will be filtered to cases where GroupLevel == "Site".

dfGroups

data.frame Group-level metadata dictionary.

dfMetrics

data.frame Metric-specific metadata for use in charts and reporting.

dfBounds

data.frame Set of predicted percentages/rates and upper- and lower-bounds across the full range of sample sizes/total exposure values for reporting.

dfAnalyticsInput

data.frame Participant-level metric data.

fnFetchParticipantData

function A function that takes a single strSubjectID argument such as "0001", and returns a list with components metadata and metric_data. metadata should be a named charact vector (or something that can be coerced to a character vector) of information about the participant, and metric_data should be a named list of data.frames, each of which contains information related to the named metric.

strTitle

character A title to display for the overall app.