Fetch Data for a Participant

This is a sample function demonstrating the type of function the user will supply to find additional data about a given participant.

Usage

sample_FetchParticipantData(strSubjectID)

Arguments

strSubjectID

character A SubjectID of an individual participant.

Value

A list with elements metadata (a list of participant metadata) and metric_data (a named list, where the names are the available metric summaries, and the values are data frames containing the details).

Examples

sample_FetchParticipantData("0001")
#> $metadata
#> $metadata$subjectID
#> [1] "0001"
#> 
#> $metadata$siteID
#> [1] "0X035"
#> 
#> $metadata$studyStartDate
#> [1] "2006-09-12"
#> 
#> $metadata$studyEndDate
#> [1] "2008-09-10"
#> 
#> $metadata$timeOnStudy
#> [1] 730
#> 
#> $metadata$country
#> [1] "US"
#> 
#> $metadata$age
#> [1] "39"
#> 
#> $metadata$sex
#> [1] "F"
#> 
#> $metadata$race
#> [1] "ASIAN"
#> 
#> 
#> $metric_data
#> $metric_data$adverseEvents
#> # A tibble: 5 × 8
#>   subjectID serious startDate  endDate    grade dictionaryDerivedTerm
#>   <chr>     <chr>   <date>     <date>     <chr> <chr>                
#> 1 0001      N       2007-05-01 2007-06-29 MILD  Blood urine present  
#> 2 0001      N       2008-08-18 NA         MILD  Nasopharyngitis      
#> 3 0001      N       2007-06-29 NA         MILD  Headache             
#> 4 0001      N       2007-08-20 2008-09-10 MILD  Arthralgia           
#> 5 0001      N       2008-07-22 2008-08-21 MILD  Pyrexia              
#> # ℹ 2 more variables: bodySystemOrOrganClass <chr>, treatmentEmergent <chr>
#> 
#> $metric_data$protocolDeviations
#> # A tibble: 1 × 4
#>   subjectID deemedImportant deviationDate category
#>   <chr>     <chr>           <date>        <chr>   
#> 1 0001      No              NA            OTHER   
#> 
#> $metric_data$studyDisposition
#> # A tibble: 0 × 2
#> # ℹ 2 variables: subjectID <chr>, studyDiscontinuationReason <chr>
#> 
#> $metric_data$treatmentDisposition
#> # A tibble: 0 × 4
#> # ℹ 4 variables: subjectID <chr>, phase <chr>, treatment <chr>,
#> #   treatmentDiscontinuationReason <chr>
#> 
#>