Participant Details server

Update Participant Details when the selected participant changes.

Usage

mod_ParticipantDetails_Server(
  id,
  fnFetchData,
  chrDomains = c("AE", "ENROLL", "LB", "PD", "SDRGCOMP", "STUDCOMP", "SUBJ", "DATACHG",
    "DATAENT", "QUERY"),
  rctv_strSubjectID
)

Arguments

id

character The id for this module, widget, or other element.

fnFetchData

function A function that takes a strDomain argument and optional strSiteID and strSubjectID and returns a data.frame. See sample_fnFetchData() for an example. The returned data.frame contains information about the named domain.

chrDomains

character A vector of names of domain data.frames to show in the app. Supported domains are shown in the default value.

rctv_strSubjectID

reactive character A shiny::reactive() object that returns the SubjectID of the selected participant.

Value

A shiny::reactive() with the ids of rows selected in the table (currently unused but implemented as a proof-of-concept for future extensions).