Run the app

run_gsm_app() creates a Shiny app to explore a set of clinical trial data. The app facilitates exploration of the data by allowing the user to click to diver deeper into aspects of the data.

run_sample_gsm_app() is a thin wrapper around run_gsm_app(), using the sample data provided in this package.

Usage

run_gsm_app(
  dfAnalyticsInput,
  dfBounds,
  dfGroups,
  dfMetrics,
  dfResults,
  fnFetchData,
  chrDomains = c("AE", "ENROLL", "LB", "PD", "SDRGCOMP", "STUDCOMP", "SUBJ", "DATACHG",
    "DATAENT", "QUERY"),
  lPlugins = NULL,
  strTitle = "GSM Deep Dive",
  strFavicon = "angles-up",
  strFaviconColor = "#FF5859",
  tagListSidebar = NULL,
  fnServer = NULL
)

run_sample_gsm_app(strFavicon = "angles-up", strFaviconColor = "#FF5859")

Arguments

dfAnalyticsInput

data.frame Participant-level metric data.

dfBounds

data.frame Set of predicted percentages/rates and upper- and lower-bounds across the full range of sample sizes/total exposure values for reporting.

dfGroups

data.frame Group-level metadata dictionary.

dfMetrics

data.frame Metric-specific metadata for use in charts and reporting.

dfResults

data.frame A stacked summary of analysis pipeline output. This will be filtered to cases where GroupLevel == "Site".

fnFetchData

function A function that takes a strDomain argument and optional strSiteID and strSubjectID and returns a data.frame. See sample_fnFetchData() for an example. The returned data.frame contains information about the named domain.

chrDomains

character A vector of names of domain data.frames to show in the app. Supported domains are shown in the default value.

lPlugins

list Optional list of plugins to include in the app.

strTitle

character A title to display for the overall app.

strFavicon

character The name of an icon to use in the browser tab via favawesome::fav().

strFaviconColor

character The hexcode or name of a color to use as the icon fill for favawesome::fav().

tagListSidebar

taglist An optional htmltools::tagList() of additional elements to add to the top of the app sidebar.

fnServer

function A Shiny server function that takes arguments input, output, and session. This function will be called at the start of the main app server function.

Value

An object that represents the app. Printing the object or passing it to shiny::runApp() runs the app.

See also

Examples

if (FALSE) { # interactive()
# Sample app will launch in a new browser window.
run_sample_gsm_app()
}