Fetch Data for a Participant

This is a sample function demonstrating the type of function the user will supply to find additional data about a given participant.

Usage

sample_FetchParticipantData(strSubjectID)

Arguments

strSubjectID

character A SubjectID of an individual participant.

Value

A list with elements metadata (a list of participant metadata) and metric_data (a named list, where the names are the available metric summaries, and the values are data frames containing the details).

See also

Examples

sample_FetchParticipantData("0008")
#> $metadata
#> $metadata$SubjectID
#> [1] "0008"
#> 
#> $metadata$enrolled
#> [1] "Y"
#> 
#> $metadata$study_start_date
#> [1] "2004-10-17"
#> 
#> $metadata$days_on_study
#> [1] 235
#> 
#> $metadata$treatment_start_date
#> [1] "2004-11-13"
#> 
#> $metadata$days_on_treatment
#> [1] 203
#> 
#> $metadata$age
#> [1] "49"
#> 
#> $metadata$sex
#> [1] "M"
#> 
#> $metadata$race
#> [1] "ASIAN"
#> 
#> $metadata$ethnicity
#> [1] "out of bound:Ethnic"
#> 
#> 
#> $metric_data
#> $metric_data$AdverseEvents
#> # A tibble: 0 × 7
#> # ℹ 7 variables: SubjectID <chr>, serious <chr>, start_date <date>,
#> #   end_date <date>, preferred_term <chr>, system_organ_class <chr>,
#> #   toxicity_grade <chr>
#> 
#> $metric_data$DataEntry
#> # A tibble: 77 × 5
#>    SubjectID data_entry_lag visit visit_date form                             
#>    <chr>              <int> <chr> <date>     <chr>                            
#>  1 0008                   5 Day 1 2004-11-13 Enrollment                       
#>  2 0008                   1 Day 1 2004-11-13 PK                               
#>  3 0008                   0 Day 1 2004-11-13 Prior and Concomitant Medication 
#>  4 0008                   5 Day 1 2004-11-13 Study Drug Accountability        
#>  5 0008                   3 Day 1 2004-11-13 Study Drug Administration (DRUG1)
#>  6 0008                   0 Day 1 2004-11-13 Study Drug Administration (DRUG2)
#>  7 0008                   7 Day 1 2004-11-13 Study Drug Administration (DRUG3)
#>  8 0008                   1 Day 1 2004-11-13 Study Drug Administration (PK)   
#>  9 0008                   8 Day 1 2004-11-13 Visit Date                       
#> 10 0008                   5 Day 1 2004-11-13 Vital Signs Performed            
#> # ℹ 67 more rows
#> 
#> $metric_data$Enrollment
#> # A tibble: 1 × 5
#>   SubjectID IntakeID        enrolled enrollment_date screen_failure_reason
#>   <chr>     <chr>           <chr>    <date>          <chr>                
#> 1 0008      X1030008-069XXX Y        2004-10-17      ""                   
#> 
#> $metric_data$Lab
#> # A tibble: 365 × 6
#>    SubjectID toxicity_grade visit     visit_date lab_test                 result
#>    <chr>     <chr>          <chr>     <date>     <chr>                     <dbl>
#>  1 0008      "1"            Screening 2004-10-17 ALT (SGPT)               66    
#>  2 0008      "2"            Screening 2004-10-17 AST (SGOT)               99    
#>  3 0008      "2"            Screening 2004-10-17 Albumin-QT               27    
#>  4 0008      "3"            Screening 2004-10-17 Alkaline Phosphatase    680    
#>  5 0008      ""             Screening 2004-10-17 Basophils                 0.100
#>  6 0008      ""             Screening 2004-10-17 Basophils (%)             1.10 
#>  7 0008      ""             Screening 2004-10-17 Calcium (EDTA)            2.06 
#>  8 0008      "0"            Screening 2004-10-17 Calcium Corrected for …   2.32 
#>  9 0008      ""             Screening 2004-10-17 Cholesterol (High Perf…   3.80 
#> 10 0008      "0"            Screening 2004-10-17 Creatine Kinase         164    
#> # ℹ 355 more rows
#> 
#> $metric_data$ProtocolDeviations
#> # A tibble: 2 × 4
#>   SubjectID deviation_date deemed_important company_category
#>   <chr>     <date>         <chr>            <chr>           
#> 1 0008      NA             No               OTHER           
#> 2 0008      NA             No               OTHER           
#> 
#> $metric_data$Queries
#> # A tibble: 10 × 6
#>    SubjectID visit                            form  field query_status query_age
#>    <chr>     <chr>                            <chr> <chr> <chr>            <dbl>
#>  1 0008      Day 1                            Vita… VSPE… Closed               6
#>  2 0008      Day 1                            Vita… WEIG… Closed              13
#>  3 0008      Early Study Drug Discontinuation Visi… Z_AG… Closed               1
#>  4 0008      Screening                        Visi… VISI… Closed               2
#>  5 0008      Week 12                          Stud… DARE… Closed               1
#>  6 0008      Week 16                          Visi… SVAE… Closed              18
#>  7 0008      Week 16                          Vita… HEIG… Closed               3
#>  8 0008      Week 20                          Visi… SVECG Closed               8
#>  9 0008      Week 4                           Prio… CMRO… Closed               2
#> 10 0008      Week 4                           Vita… VSPE… Closed              11
#> 
#> $metric_data$StudyCompletion
#> # A tibble: 1 × 3
#>   SubjectID completed_study discontinuation_reason
#>   <chr>     <chr>           <chr>                 
#> 1 0008      N               WITHDRAWAL BY SUBJECT 
#> 
#> $metric_data$TreatmentCompletion
#> # A tibble: 1 × 4
#>   SubjectID phase                     completed_treatment discontinuation_reason
#>   <chr>     <chr>                     <chr>               <chr>                 
#> 1 0008      Blinded Study Drug Compl… N                   out of bound:Withdrew…
#> 
#>