Parameters used in multiple functions

Reused parameter definitions are gathered here for easier usage. Note: We intentionally use id rather than the more standardized strID throughout this package, because shiny::testServer() specifically expects an id argument in module server functions.

Arguments

chrAllowedFields

character A vector of names of allowed fields in an object.

chrAllowedValues

character A vector of allowed values for a vector.

chrDomains

character A vector of names of domain data.frames to show in the app. Supported domains are shown in the default value.

chrFromTabs

character A vector of tab names. Syncing is only triggered if the user is currently on one of these tabs.

chrLinkIDs

character Module ids for multiple modules.

chrLabels

character A vector of labels for display to the user.

chrMessage

character A vector of message elements to be formatted via cli::cli_bullets().

chrMetrics

character A named vector of MetricIDs, where the names are the full Metric.

chrPluginFiles

character The files in a plugin directory.

chrRequiredColumns

character A vector of expected columns in a data.frame.

chrRequiredFields

character A vector of names of required fields in an object.

chrRequiredInputs

character A vector of inputs required by a Plugin.

chrSites

character A vector of sites available in the study.

chrValues

character A vector of values to associate with a vector of labels.

dfAnalyticsInput

data.frame Participant-level metric data.

dfBounds

data.frame Set of predicted percentages/rates and upper- and lower-bounds across the full range of sample sizes/total exposure values for reporting.

dfGroups

data.frame Group-level metadata dictionary.

dfMetrics

data.frame Metric-specific metadata for use in charts and reporting.

dfParticipantGroups

data.frame Unique SubjectID and GroupID combos from dfAnalyticsInput.

dfResults

data.frame A stacked summary of analysis pipeline output. This will be filtered to cases where GroupLevel == "Site".

envCall

environment The environment from which this function was called, for use in better error messages. This value should usually be left as the default, or passed from the calling function if the calling function also has an envCall argument.

envEvaluate

environment The environment in which any variables used in the message are defined. You almost definitely want to leave this as the default value.

fnFetchData

function A function that takes a strDomain argument and optional strSiteID and strSubjectID and returns a data.frame. See sample_fnFetchData() for an example. The returned data.frame contains information about the named domain. If the function throws an error, the error is elevated to the user, so you can use errors to pass requirements through to the user.

fnServer

function A Shiny server function that takes arguments input, output, and session. This function will be called at the start of the main app server function.

id

character The id for this element.

intKRIColorCount

integer A named vector of counts by color.

intNParticipants

integer The number of unique participants (subjects) in the study.

intRed

integer The number of sites with at least one red flag.

intAmber

integer The number of sites with at least one amber flag.

lMetric

list Named list of data describing a single metric, as well as things like which group is selected.

l_rctvDomains

list A named list of shiny::reactive() objects, each of which returns a domain dataframe.

lPluginDefinition

list A named list with required elements strTitle, fnUI, and fnServer, and optional fields packages and lConfig. Usually generated by plugin_Read().

lPlugins

list Optional list of plugins to include in the app.

lStudy

list Named list of data describing the overall study.

rctv_dfBounds

reactive dataframe A shiny::reactive() object that returns a set of predicted percentages/rates and upper- and lower-bounds across the full range of sample sizes/total exposure values for reporting.

rctv_dfData

reactive dataframe A shiny::reactive() object that returns a generic dataframe.

rctv_dfDomain

reactive dataframe A shiny::reactive() object that returns a domain dataframe.

rctv_dfResults

reactive dataframe A shiny::reactive() object that returns a stacked summary of analysis pipeline output.

rctv_gtObject

reactive gt_table A shiny::reactive() object that returns a gt::gt() object.

rctv_lMetric

reactive list A shiny::reactive() object that returns a named list of data describing a single metric, as well as things like which group is selected.

rctv_lglState

reactive logical A shiny::reactive()] object that returns a Boolean value indicating whether something is "off" (FALSE) or "on" (TRUE).

rctv_strDomainID

reactive character A shiny::reactive() object that returns the selected DomainID (such as "AE" or "SUBJ").

rctv_strCurrentTab

reactive character A shiny::reactive() object that returns the currently selected tab.

rctv_strGroupSubset

reactive character A shiny::reactive() object that returns the selected subset of groups to include in the table.

rctv_strInput

reactive character A shiny::reactive() object that returns the value(s) of an input to share with an interactive gt table. If this argument is a shiny::reactiveVal(), it will be used to push updates back to the calling function.

rctv_strMetricID

reactive character A shiny::reactive() object that returns the selected MetricID.

rctv_strName

reactive character A shiny::reactive() object that returns the name of an object, such as a particular dataframe in a named list.

rctv_strSiteID

reactive character A shiny::reactiveVal() object that returns the GroupID of the selected site, and can be used to update the selected site.

rctv_strSubjectID

reactive character A shiny::reactive() object that returns the SubjectID of the selected participant.

rctv_strValue

reactive character A shiny::reactive() object that returns the value of a field.

session

environment Session from which to make a child scope (the default should almost always be used).

strArg

character The name of the argument in the calling function. In general, this value should either be left as the default, or passed from the calling function if the calling function also has a strArg argument.

strClass

character A descriptive label for this type of error, in lower_snake_case.

strColorFamily

character Whether to load the "dark" version of this color, or the "light" version. Default: "dark".

strColorName

character The target color to report about.

strColorCode

character The hex code (such as "#FFFFFF") for a color.

strContainerID

character The (namespaced) ID of the target container (usually a div).

strDomain

character The domain data.frame to load. Supported domains are shown in the default value.

strEmpty

character The value to return when everything is deselected.

strFavicon

character The name of an icon to use in the browser tab via favawesome::fav().

strFaviconColor

character The hexcode or name of a color to use as the icon fill for favawesome::fav().

strFunction

character The name of a function, with or without namespace ("reactive" or "shiny::reactive"). If a namespace is supplied, we check that the package is installed, and attach it. The function can also be supplied as a function or a purrr-style formula. See rlang::as_function() (the x argument) for more details.

strGroupID

character A GroupID to focus on.

strGroupLabelKey

character Value for the group label key. Default: "InvestigatorLastName".

strGroupLevel

character Value for the group level. Default: NULL and taken from dfMetrics$GroupLevel if available.

strGroupSubset

character Subset of groups to include in the table. Default: "red". Options:

"all": All groups.

"red": Groups with 1+ red flags.

"red/amber": Groups with 1+ red/amber flag.

"amber": Groups with 1+ amber flag.

strInputID

character An ID to use for the Shiny input created by this module or used by this JavaScript.

strInputName

character The name of an input. One of "site", "participant", or "domain".

strLabel

character The label of a field.

strMetricID

character A MetricID to focus on.

strOutcome

character Outcome variable. Default: "Score".

strPlotTitle

character A title for a plot, usually the name of a metric.

strSiteID

character A GroupID of an individual site within a study.

strSubjectID

character A SubjectID of an individual participant.

strTargetTab

character The tab to switch to.

strText

character Text to display.

strTitle

character A title to display for the overall app.

strValue

character The value of a field.

strWhat

character A sentence-case description of the object being inspected.

strWidgetName

character The name of a widget in the gsm package.

tagListSidebar

taglist An optional htmltools::tagList() of additional elements to add to the top of the app sidebar.

x

An object to validate.